# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.adminSetLocationRequest:
    properties:
      date:
        type: string
      location:
        type: string
      reason:
        type: string
      user_id:
        type: string
    required:
    - date
    - location
    - user_id
    type: object
...
  handlers.setLocationRequest:
    properties:
      date:
        type: string
      location:
        type: string
    required:
    - date
    - location
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@craftsbite.com
    name: API Support
  description: CraftsBite Backend API for meal management system
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: CraftsBite API
  version: "1.0"
paths:
...
  /work-location:
    get:
      description: Get work location for current user on a specific date
      parameters:
      - description: Date (YYYY-MM-DD)
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Work location retrieved
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get my work location
      tags:
      - work-location
    post:
      consumes:
      - application/json
      description: Set work location (office/wfh) for current user on a specific date
      parameters:
      - description: Location details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.setLocationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Work location updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Set my work location
      tags:
      - work-location
  /work-location/list:
    get:
      description: Get all users' work locations for a specific date (Admin/Team Lead
        only)
      parameters:
      - description: Date (YYYY-MM-DD)
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Work locations retrieved
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List work locations by date
      tags:
      - work-location
  /work-location/monthly-summary:
    get:
      description: Get monthly work-from-home summary for current user
      parameters:
      - description: Month (YYYY-MM, defaults to current month)
        in: query
        name: month
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Monthly WFH summary retrieved
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get monthly WFH summary
      tags:
      - work-location
  /work-location/override:
    post:
      consumes:
      - application/json
      description: Set work location for another user (Admin/Team Lead only)
      parameters:
      - description: Override details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.adminSetLocationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Work location corrected successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Override work location
      tags:
      - work-location
  /work-location/team-monthly-report:
    get:
      description: Get monthly work-from-home report for team (Admin/Logistics/Team
        Lead only)
      parameters:
      - description: Month (YYYY-MM, defaults to current month)
        in: query
        name: month
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Monthly WFH report retrieved
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get team monthly WFH report
      tags:
      - work-location
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"